--- a/stoplight.gemspec
+++ b/stoplight.gemspec
@@ -21,6 +21,8 @@ Gem::Specification.new do |gem|
 
   gem.required_ruby_version = '>= 1.9.3'
 
+  gem.add_dependency 'multi_json', '~> 1.10'
+
   gem.add_development_dependency 'benchmark-ips', '~> 2.1'
   gem.add_development_dependency 'coveralls', '~> 0.7'
   gem.add_development_dependency 'fakeredis', '~> 0.5'