Fixes #65. 

This pull request adds a dependency on MultiJson. That allows JSON encoding and decoding to use faster gems like Oj or YAJL. 

I wondered how much of an impact this would have on performance, so I ran our Redis benchmark. 

``` sh
$ bundle exec ruby bench/redis_bench.rb  
```

The results were inconclusive. Here's how it looked at first:

```
Calculating -------------------------------------
     creating lambda    65.948k i/100ms
      calling lambda    58.882k i/100ms
  creating stoplight    35.259k i/100ms
   calling stoplight   837.000  i/100ms
-------------------------------------------------
     creating lambda      1.244M (±16.5%) i/s -      6.067M
      calling lambda      1.130M (±13.2%) i/s -      5.535M
  creating stoplight    466.061k (±12.4%) i/s -      2.327M
   calling stoplight      8.598k (± 7.4%) i/s -     43.524k

Comparison:
     creating lambda:  1243760.1 i/s
      calling lambda:  1129788.3 i/s - 1.10x slower
  creating stoplight:   466060.5 i/s - 2.67x slower
   calling stoplight:     8598.5 i/s - 144.65x slower
```

And here's how it looked with Oj 2.11.4:

```
Calculating -------------------------------------
     creating lambda    66.011k i/100ms
      calling lambda    62.826k i/100ms
  creating stoplight    35.500k i/100ms
   calling stoplight   812.000  i/100ms
-------------------------------------------------
     creating lambda      1.243M (±15.5%) i/s -      6.073M
      calling lambda      1.148M (±13.8%) i/s -      5.654M
  creating stoplight    475.940k (±12.9%) i/s -      2.343M
   calling stoplight      8.505k (± 6.9%) i/s -     43.036k

Comparison:
     creating lambda:  1243141.7 i/s
      calling lambda:  1148103.4 i/s - 1.08x slower
  creating stoplight:   475939.9 i/s - 2.61x slower
   calling stoplight:     8505.2 i/s - 146.16x slower
```

From those results, I can only assume that JSON encoding and decoding represents an insignificant part of our runtime. I'm not sure it's worth adding the dependency, but I'd be interested to hear what @AaronLasseigne and @camdez think.