@tfausak 2015-02-08
I realized the benchmarks don't even touch JSON because they don't make any failures. I'll have to make a benchmark that does cause some failures.