@tfausak 2015-02-08
I made a more pathological benchmark:

``` rb
require 'benchmark/ips'
require 'fakeredis'
require 'stoplight'

redis = Redis.new
data_store = Stoplight::DataStore::Redis.new(redis)
Stoplight::Light.default_data_store = data_store

Benchmark.ips do |b|
  b.report('failing') do
    begin
      fail
    rescue
      nil
    end
  end
  b.report('failing stoplight') do
    begin
      Stoplight::Light.new('') { fail }.with_threshold(Float::INFINITY).run
    rescue
      nil
    end
  end
  b.compare!
end
```

Without Oj:

```
Calculating -------------------------------------
             failing    39.034k i/100ms
   failing stoplight    13.000  i/100ms
-------------------------------------------------
             failing    565.487k (± 6.4%) i/s -      2.849M
   failing stoplight     48.892  (±18.4%) i/s -    247.000 

Comparison:
             failing:   565486.6 i/s
   failing stoplight:       48.9 i/s - 11566.01x slower
```

With Oj:

```
Calculating -------------------------------------
             failing    39.519k i/100ms
   failing stoplight    13.000  i/100ms
-------------------------------------------------
             failing    569.049k (± 7.3%) i/s -      2.845M
   failing stoplight     50.623  (±19.8%) i/s -    260.000  in   5.325417s

Comparison:
             failing:   569049.0 i/s
   failing stoplight:       50.6 i/s - 11240.96x slower
```

There's still not much of a difference.